Fixed get_network_traffic method to return the pcap data
Calling _return_response_and_status_code in get_network_traffic will fail
because the response will be a binary blob (the pcap data) and not a json
object. Therefore calling response.json() will throw a ValueError exception.

Instead, we can catch this exception and return the raw binary data (the pcap).
If VirusTotal is unable to retrieve the pcap data, it will return a json object
as usual.

The Python magic module will only return a string describing the file type - it
will not actually return the raw pcap data.